+++ docs/release-notes.md
@@ -18,6 +18,7 @@
 
 * For modders:
   * Added asset propagation for `Data\Concessions`.
+  * Improved error-handling during asset propagation.
 
 ## 3.9.5
 Released 21 March 2021 for Stardew Valley 1.5.4 or later.